Added regtech-regex in pyproject, updated code to use (#138)
Closes #135 

Thought about adding a check for rssd_id but by saying it has to be an
int in the DTO there's no need for the check. Only negative (hesitate to
call it that) is that we wouldn't be using the standard regtech-regex
error_text. But that's probably fine, unless we want to change rssd_id
to a string and then check the regex.